feat(utils): add file existence check before conversion to hcpe
Add a check in plaintext_to_hcpe.py to skip the conversion of files to .hcpe format if the output file already exists. This prevents overwriting existing files and unnecessary reprocessing.

In hcpe_to_hdf5.sh, update the input file list by adding 'suisho10Mn.hcpe' to the dataset for shuffling. This change expands the data used during the shuffling process, potentially improving the diversity and quality of the training data set.