refactor(dlshogi): change export format from ONNX to NPZ and clean up code
This commit changes the model serialization format from ONNX to NPZ for the dlshogi project. The script `wandb_checkpoint_to_onnx.py` is renamed to `wandb_checkpoint_to_model.py`. The ONNX-specific code has been removed, and the script now uses the `serializers` module to save models in NPZ format. Additionally, the command-line arguments have been updated to reflect these changes, with the addition of new arguments for specifying the artifact path and download path. This simplification results in a cleaner and more focused script for model serialization.